+++ src/components/Dialog/Dialog.js
@@ -132,6 +132,7 @@ export const Dialog = ({
   );
 };
 
+Dialog.propTypes = {
   /**
    * Title of the dialog
    */