+++ src/components/V2Button/V2Button.js
@@ -4,47 +4,53 @@ import { string, bool, node, func, oneOf } from 'prop-types';
 import classnames from 'classnames';
 import * as olt from '@lightelligence/styles';
 
-const V2Button = ({
-  className,
-  emphasis,
-  // avoid name clash with type which can be passed to input elements
-  buttonType,
-  theme,
-  children,
-  disabled,
-  tag,
-  icon,
-  iconLeft,
-  iconRight,
-  onClick,
-  ...props
-}) => {
-  const Element = tag || 'button';
-  const isIconDefault = icon && !emphasis && !buttonType;
-  return (
-    <Element
-      disabled={disabled}
-      onClick={onClick}
-      {...props}
-      className={classnames(
-        olt.V2Button,
-        emphasis && olt[`V2Button${pascalize(emphasis)}`],
-        buttonType && olt[`V2Button${pascalize(buttonType)}`],
-        theme && olt[`V2Button${pascalize(theme)}`],
-        isIconDefault && olt.V2ButtonIconDefault,
-        icon && olt.V2ButtonIcon,
-        icon && olt[`Icon${pascalize(icon)}`],
-        iconLeft && olt[`Icon${pascalize(iconLeft)}`],
-        iconLeft && olt.V2ButtonIconLeft,
-        iconRight && olt[`Icon${pascalize(iconRight)}`],
-        iconRight && olt.V2ButtonIconRight,
-        className,
-      )}
-    >
-      {children}
-    </Element>
-  );
-};
+const V2Button = React.forwardRef(
+  (
+    {
+      className,
+      emphasis,
+      // avoid name clash with type which can be passed to input elements
+      buttonType,
+      theme,
+      children,
+      disabled,
+      tag,
+      icon,
+      iconLeft,
+      iconRight,
+      onClick,
+      ...props
+    },
+    ref,
+  ) => {
+    const Element = tag || 'button';
+    const isIconDefault = icon && !emphasis && !buttonType;
+    return (
+      <Element
+        disabled={disabled}
+        onClick={onClick}
+        ref={ref}
+        {...props}
+        className={classnames(
+          olt.V2Button,
+          emphasis && olt[`V2Button${pascalize(emphasis)}`],
+          buttonType && olt[`V2Button${pascalize(buttonType)}`],
+          theme && olt[`V2Button${pascalize(theme)}`],
+          isIconDefault && olt.V2ButtonIconDefault,
+          icon && olt.V2ButtonIcon,
+          icon && olt[`Icon${pascalize(icon)}`],
+          iconLeft && olt[`Icon${pascalize(iconLeft)}`],
+          iconLeft && olt.V2ButtonIconLeft,
+          iconRight && olt[`Icon${pascalize(iconRight)}`],
+          iconRight && olt.V2ButtonIconRight,
+          className,
+        )}
+      >
+        {children}
+      </Element>
+    );
+  },
+);
 
 V2Button.propTypes = {
   /**