+++ CHANGELOG.md
@@ -38,3 +38,4 @@
 - Fix: className property for Tabs Component
 - Fix: Design for Notification
 - Fix: Docs for v2 Modal
+